feat(logger): add logger utility module with console logging functions for info, success and error levels feat(logger): add logger enum module with console color codes for foreground and background colors and effects
